+++ packages/hydrogen/src/components/LocalizationProvider/LocalizationContext.client.tsx
@@ -9,16 +9,5 @@ export interface LocalizationContextValue {
   setCountry(country: Localization['country']): void;
 }
 
-/**
- * This is a workaround to support exposing context from within a server component
- * exported from Hydrogen.
- */
-declare global {
-  var __localizationContext: any;
-}
-
 export const LocalizationContext =
-  globalThis.__localizationContext ||
   createContext<LocalizationContextValue | null>(null);
-
-globalThis.__localizationContext = LocalizationContext;