+++ packages/playground/server-components/src/components/Link.client.jsx
@@ -1,1 +1,0 @@
-export {Link} from 'react-router-dom';