+++ packages/hydrogen/CHANGELOG.md
@@ -8,6 +8,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ## Unreleased
 
 - fix: reading property of null for component props
+- fix: gallery safari grid layout
 
 ## 0.6.1 - 2021-11-08