+++ packages/hydrogen/CHANGELOG.md
@@ -7,12 +7,16 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+- fix: remove CartProvider from BuyNowButton
 - fix: reading property of null for component props
-- fix: gallery safari grid layout
+- fix: transform deeply-imported client components
+- fix: duplicated files and contexts in browser
 
 ## 0.6.1 - 2021-11-08
 
 - No updates. Transitive dependency bump.
+
+- fix: do not set headers after they are sent to client
 
 ## 0.6.0 - 2021-11-05